Update RouteRetrievalEvent (#1731)

+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/NavigationMetricsWrapper.java
@@ -371,7 +371,7 @@ final class NavigationMetricsWrapper {
     mapboxTelemetry.push(feedbackEvent);
   }
 
-  static void routeRetrievalEvent(long elapsedTime, String routeUuid, String sessionId) {
+  static void routeRetrievalEvent(double elapsedTime, String routeUuid, String sessionId) {
     push(new RouteRetrievalEvent(elapsedTime, routeUuid, sessionId));
   }
